requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-docx
pandas
pdfplumber
python-multipart
openpyxl
google-generativeai
python-dotenv
